Move modules file to modules folder
