+++ src/index.js
@@ -1,6 +1,6 @@
 import './style.css';
 import logoBack from './assets/cookBanner3.jpg';
-import getAllData from './getDataFood.js';
+import getAllData from './modules/getDataFood.js';
 
 // eslint-disable-next-line operator-linebreak
 const [liBeefElement, liPastaElement, liSeafoodElement] =
